backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
Pillow==11.2.1
python-dotenv==1.1.0
